XML Exercise - Problem 19 updated

+++ XML/CarDealer/DTOs/Export/ExportCarSalesWithDiscNondiscPrices.cs
@@ -26,7 +26,7 @@ namespace CarDealer.DTOs.Export
 
 
         [XmlElement("price-with-discount")] 
-        public decimal PriceDiscounted { get; set; }
+        public double PriceDiscounted { get; set; }
 
 
     }
